Add extra_args for zoxide query command.
This allows me to call `switch_workspace(" | rg -Fxf ~/.projects")` to filter the list
further to only those candidates I want to include in the selection, while maintaining
the query ordering behavior benefits of using zoxide in the first place. But this could
be used for any additional argument somebody wanted to use for the query. Of course, it
is likely none or few will find this useful, but it is unobtrusive. If you reject it, I
completely understand, but this is useful for me.

As an aside, if you're interested in how I generate this list, I use [this script](https://github.com/theherk/commons/blob/main/bin/repocache.sh)
to find all git roots in my projects directory and populate a file that is just a list
of these for filtering.